Added discount price cart && agree modals

+++ src/providers/AuthProvider.tsx
@@ -8,6 +8,8 @@ import {
   USER_CUSTOMER,
 } from '../contstants/storage-keys.constants.ts';
 import { localStorageService } from '../services';
+import {useModal} from "../hooks/useModal.ts";
+import {useNavigate} from "react-router-dom";
 
 export interface IAuthContext {
   user: Customer | null;
@@ -27,6 +29,8 @@ export const AuthProvider = ({ children }: Props) => {
   const [isAuthenticated, setIsAuthenticated] = useState(!!customer);
   const [user, setUser] = useState<Customer | null>(customer);
   const { enqueueSnackbar } = useSnackbar();
+  const { openModal, closeModal } = useModal();
+  const navigate = useNavigate();
 
   const login = (user: Customer) => {
     localStorageService.set<Customer>(USER_CUSTOMER, user);
@@ -35,13 +39,24 @@ export const AuthProvider = ({ children }: Props) => {
   };
 
   const logout = () => {
-    localStorageService.remove(ACCESS_TOKEN);
-    localStorageService.remove(REFRESH_TOKEN);
-    localStorageService.remove(EXPIRATION_TIME);
-    localStorageService.remove(USER_CUSTOMER);
-    setUser(null);
-    setIsAuthenticated(false);
-    enqueueSnackbar('Вы вышли из системы', { variant: 'success' });
+    openModal({
+      open: true,
+      title: 'Выйти из учетной записи?',
+      description: '',
+      handleClose: (agree: boolean) => {
+        closeModal();
+        if (agree) {
+          localStorageService.remove(ACCESS_TOKEN);
+          localStorageService.remove(REFRESH_TOKEN);
+          localStorageService.remove(EXPIRATION_TIME);
+          localStorageService.remove(USER_CUSTOMER);
+          setUser(null);
+          setIsAuthenticated(false);
+          navigate('/login');
+          enqueueSnackbar('Вы вышли из системы', { variant: 'success' });
+        }
+      }
+    })
   };
 
   return (
